+++ package.json
@@ -43,31 +43,35 @@
 		"cmd",
 		"console"
 	],
-	"dependencies": {
+	"dependencies": {},
+	"devDependencies": {
+		"@babel/core": "^7.2.2",
+		"@babel/preset-env": "^7.2.3",
+		"@babel/register": "^7.0.0",
+		"ava": "^1.0.1",
 		"camelcase-keys": "^5.0.0",
-		"decamelize-keys": "^1.0.0",
+		"decamelize-keys": "^1.1.0",
+		"esm": "^3.0.84",
+		"execa": "^1.0.0",
 		"hard-rejection": "^1.0.0",
-		"minimist-options": "^3.0.1",
-		"normalize-package-data": "^2.3.4",
+		"indent-string": "^3.2.0",
+		"minimist-options": "^3.0.2",
+		"normalize-package-data": "^2.4.0",
 		"read-pkg-up": "^4.0.0",
 		"redent": "^2.0.0",
-		"trim-newlines": "^2.0.0",
-		"yargs-parser": "^11.0.0"
-	},
-	"devDependencies": {
-		"ava": "^0.25.0",
-		"execa": "^1.0.0",
-		"indent-string": "^3.0.0",
-		"rollup": "^0.66.2",
-		"rollup-config": "^1.0.1",
-		"rollup-plugin-commonjs": "^9.1.8",
+		"repl": "^0.1.3",
+		"rollup": "^1.0.1",
+		"rollup-plugin-commonjs": "^9.2.0",
 		"rollup-plugin-json": "^3.1.0",
-		"rollup-plugin-node-resolve": "^3.4.0",
-		"xo": "^0.23.0"
+		"rollup-plugin-node-resolve": "^4.0.0",
+		"trim-newlines": "^2.0.0",
+		"xo": "^0.23.0",
+		"yargs-parser": "^11.1.1"
 	},
 	"xo": {
 		"rules": {
-			"unicorn/no-process-exit": "off"
+			"unicorn/no-process-exit": "off",
+			"ava/no-ignored-test-files": "warn"
 		}
 	}
 }
