+++ test/fixtures/dist.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
-'use strict'
+'use strict';
 
-const meow = require('../../dist')
+const meow = require('../../dist');
 
 const cli = meow({
 	description: 'Custom description',
@@ -16,12 +16,12 @@ const cli = meow({
 		meow: {default: 'dog'},
 		camelCaseOption: {default: 'foo'}
 	}
-})
+});
 
 if (cli.flags.camelCaseOption === 'foo') {
 	Object.keys(cli.flags).forEach(x => {
-		console.log(x)
-	})
+		console.log(x);
+	});
 } else {
-	console.log(cli.flags.camelCaseOption)
+	console.log(cli.flags.camelCaseOption);
 }
